21-ghost-rf-fingerprinting/requirements.txt
streamlit>=1.30
folium>=0.15
streamlit-folium>=0.18
plotly>=5.20
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
openai>=1.40
python-dotenv>=1.0
Pillow>=10.0
